change assets/everlost/ to assetPacks/test/ and -remove- per-class failsafes for trying to get null textures from warehouse
yes, removed the failsafes. warehouse was -supposed- to handle if a null texture was asked for by returning the missing texture, but texture remained null because warehouse was never asked.

also added some notes about my concerns as to how i should handle everlost isle's future development: move to this engine completely, or develop this simultaneously/for fun?